+++ packages/plugin-autocapture-browser/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.28.9](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-autocapture-browser@1.28.8...@amplitude/plugin-autocapture-browser@1.28.9) (2026-07-22)
+
+**Note:** Version bump only for package @amplitude/plugin-autocapture-browser
+
+
+
+
+
+## [1.28.8](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-autocapture-browser@1.28.7...@amplitude/plugin-autocapture-browser@1.28.8) (2026-07-20)
+
+
+### Bug Fixes
+
+* **element-selector:** resolve config race condition ([#1892](https://github.com/amplitude/Amplitude-TypeScript/issues/1892)) ([8f9a310](https://github.com/amplitude/Amplitude-TypeScript/commit/8f9a310f5e41ed69647cfff0300db17677e2c8de))
+
+
+
+
+
 ## [1.28.7](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-autocapture-browser@1.28.6...@amplitude/plugin-autocapture-browser@1.28.7) (2026-07-17)
 
 **Note:** Version bump only for package @amplitude/plugin-autocapture-browser